version estable pero mala

+++ include/Theo.cuh
@@ -1,8 +1,8 @@
 __host__ void FillSigma ();
 __host__ void FillEnergy ();
-__host__ float Sigma (float r);
-__host__ float Energy (float r);
-__host__ float CoolingTime (float r);
+__host__ float Sigma (double r);
+__host__ float Energy (double r);
+__host__ float CoolingTime (double r);
 __host__ void FillCoolingTime ();
-__host__ float Qplusinit (float r);
+__host__ float Qplusinit (double r);
 __host__ void FillQplus ();
